pyproject.toml
[tool.poetry]
name = "nexus"
version = "0.1.0"
description = "Project Nexus AI Foundation"
authors = ["Vibhor Kashmira"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.34"}
pydantic = "^2.10"
pydantic-settings = "^2.7"
qdrant-client = "^1.12"
supabase = "^2.13"
upstash-redis = "^1.2"
langfuse = "^2.55"
python-dotenv = "^1.0"
unstructured = {extras = ["pdf"], version = "^0.15"}
simhash = "^2.0"
sentence-transformers = "^3.0"
tiktoken = "^0.7"
spacy = "^3.7"
yake = "^0.4"
ragas = "0.1.7"
datasets = "^2.19"
pandas = "^2.2"
langchain-openai = "^0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
pytest-asyncio = "^0.23"
httpx = "^0.27"
ruff = "^0.4"

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
filterwarnings = ["ignore::DeprecationWarning", "ignore::UserWarning"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
